services/api_crud/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
click==8.1.7
colorama==0.4.6
databases==0.8.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.109.0
frozenlist==1.4.1
GeoAlchemy2==0.14.3
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
multidict==6.0.4
packaging==23.2
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4